Add Permision Logit to Deleting and Update Post

+++ Project1-3/controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator/check');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -49,19 +50,28 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path.replace('\\', '/');
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: 'kooks7' }
+    creator: req.userId
   });
   post
     .save()
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: 'Post created successfully!',
-        post: result
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
     .catch(err => {
@@ -121,6 +131,12 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw err;
       }
+      // post DB와 토큰에서 가져온 userId 값 비교
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not Authorized!');
+        err.statusCode = 403;
+        throw error;
+      }
 
       // 기존 이미지 지우기
       if (imageUrl !== post.imageUrl) {
@@ -151,14 +167,27 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw err;
       }
+      // post DB와 토큰에서 가져온 userId 값 비교
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not Authorized!');
+        err.statusCode = 403;
+        throw error;
+      }
       // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: 'Delete Post.' });
     })
+
     .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
